Mv `map_cells_to_dict` `get_value` 2 cis_helper.py
