+++ cis_helper.py
@@ -1,0 +1,15 @@
+def map_cells_to_dict(cells, dic={}):
+
+    for c in cells:
+        dic[c.id] = c
+
+    return dic
+
+
+def get_value(dict, key):
+
+    try:
+        value = dict[key]
+        return value, True
+    except BaseException:
+        return None, False
